Fixed driver and tests to work in Node 18
